# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Open Agents HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/browser/commands:
    post:
      operationId: executeBrowserCommand
      parameters:
      - description: Opaque browser capability injected into the owning Open Agents
          worker.
        in: header
        name: X-OPEN-AGENTS-Browser-Capability
        schema:
          description: Opaque browser capability injected into the owning Open Agents
            worker.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrowserCommandRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowserCommandResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Service Unavailable
      summary: Execute a target-scoped command in a session's desktop browser
      tags:
      - browser
  /api/v1/browser/status:
    get:
      operationId: getBrowserStatus
      parameters:
      - description: Open Agents session identifier.
        in: query
        name: sessionId
        schema:
          description: Open Agents session identifier.
          type: string
      - description: Opaque browser capability injected into the owning Open Agents
          worker.
        in: header
        name: X-OPEN-AGENTS-Browser-Capability
        schema:
          description: Opaque browser capability injected into the owning Open Agents
            worker.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowserStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Check whether the desktop browser runtime is connected for a session
      tags:
      - browser
...
  /api/v1/dev/import-projects:
    post:
      operationId: runDevImportProjects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevImportProjectsRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevImportProjectsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Run the developer project-registry import through the daemon store
      tags:
      - dev
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    BrowserCommandRequest:
      properties:
        action:
          type: string
        args:
          additionalProperties: {}
          type: object
        sessionId:
          type: string
      required:
      - sessionId
      - action
      type: object
    BrowserCommandResponse:
      properties:
        action:
          type: string
        requestId:
          type: string
        result: {}
        sessionId:
          type: string
      required:
      - requestId
      - sessionId
      - action
      - result
      type: object
    BrowserStatusResponse:
      properties:
        connected:
          type: boolean
        connectedAt:
          format: date-time
          type: string
        sessionId:
          type: string
        transport:
          type: string
      required:
      - sessionId
      - connected
      - transport
      type: object
...
    DevImportProjectsConflict:
      properties:
        path:
          type: string
        projectId:
          type: string
        reason:
          type: string
        targetId:
          type: string
        targetPath:
          type: string
      required:
      - projectId
      - path
      - reason
      type: object
    DevImportProjectsReport:
      properties:
        conflicts:
          items:
            $ref: '#/components/schemas/DevImportProjectsConflict'
          type: array
        dryRun:
          type: boolean
        inserted:
          type: integer
        skipped:
          type: integer
        sourceDataDir:
          type: string
        targetDataDir:
          type: string
        updated:
          type: integer
      required:
      - sourceDataDir
      - targetDataDir
      - dryRun
      - inserted
      - updated
      - skipped
      type: object
    DevImportProjectsRequest:
      properties:
        dryRun:
          type: boolean
        sourceDataDir:
          minLength: 1
          type: string
      required:
      - sourceDataDir
      - dryRun
      type: object
    DevImportProjectsResponse:
      properties:
        report:
          $ref: '#/components/schemas/DevImportProjectsReport'
      required:
      - report
      type: object
...