# backend/internal/httpd/controllers/dev.go
package controllers

import (
	"context"
	"encoding/json"
	"net/http"

	"github.com/go-chi/chi/v5"

	"github.com/sudo-adduser-jordan/open-agents/backend/internal/devimport"
	"github.com/sudo-adduser-jordan/open-agents/backend/internal/httpd/apispec"
	"github.com/sudo-adduser-jordan/open-agents/backend/internal/httpd/envelope"
	devimportsvc "github.com/sudo-adduser-jordan/open-agents/backend/internal/service/devimport"
)

// DevImportService is the controller-facing developer import service contract.
type DevImportService interface {
	RunProjects(ctx context.Context, in devimportsvc.RunInput) (devimport.Report, error)
}

// DevController owns developer-only API routes.
type DevController struct {
	Import DevImportService
}

// Register mounts developer REST routes on the supplied router.
func (c *DevController) Register(r chi.Router) {
	r.Post("/dev/import-projects", c.importProjects)
}

func (c *DevController) importProjects(w http.ResponseWriter, r *http.Request) {
	if c.Import == nil {
		apispec.NotImplemented(w, r, "POST", "/api/v1/dev/import-projects")
		return
	}
	var req DevImportProjectsRequest
	dec := json.NewDecoder(r.Body)
	dec.DisallowUnknownFields()
	if err := dec.Decode(&req); err != nil {
		envelope.WriteAPIError(w, r, http.StatusBadRequest, "bad_request", "INVALID_JSON", "request body must be valid JSON", nil)
		return
	}
	if req.SourceDataDir == "" {
		envelope.WriteAPIError(w, r, http.StatusBadRequest, "bad_request", "SOURCE_DATA_DIR_REQUIRED", "sourceDataDir is required", nil)
		return
	}
	rep, err := c.Import.RunProjects(r.Context(), devimportsvc.RunInput{
		SourceDataDir: req.SourceDataDir,
		DryRun:        req.DryRun,
	})
	if err != nil {
		envelope.WriteError(w, r, err)
		return
	}
	envelope.WriteJSON(w, http.StatusOK, DevImportProjectsResponse{Report: rep})
}